DestroysGroupMemberships action implemented
